Pod firmware: PODFW-REV1 fix wave (3 HIGH + 7 MEDIUM + 2 LOW)
- The u-center bridge parses host->GNSS UBX frames and drops UBX-CFG-OTP
  (06 41), non-RAM/BBR/Flash VALSET layers and embedded OTP; a stray 0xB5
  can never splice a frame. Tested with the Table-3 bytes in every chunk
  size.
- OTP state is SET/NOT_SET only from a complete, strictly validated VALGET
  reply correlated with our poll; everything else is UNKNOWN (refuse).
- Console never blocks the loop: non-blocking writes, drop-and-count,
  non-zero TX timeout (HWCDC timeout-0 underflow confirmed in source).
  Module named console_io (CON is a reserved device name on Windows).
- BLE controls tagged by connection generation and cleared on disconnect;
  INFO served from a loop-published snapshot; bounded control drain with
  explicit BUSY; atomic rate change (one VALSET + strict readback +
  rollback); strict WiFi/GNSS coexistence verdict; PPS/IMU ISRs on the
  ESP-IDF IRAM ISR service; overflowed/malformed console lines rejected
  whole; OTP authorisation expires wrap-safely.
88 native tests pass; ESP32 build clean.

+++ firmware-pod/src/pod_protocol.c
@@ -296,7 +296,7 @@ pod_dec_result_t pod_decode_status(const pod_frame_t *f, pod_status_t *s) {
 
 pod_dec_result_t pod_decode_control(const pod_frame_t *f, pod_control_t *c) {
   if (f->type != POD_TYPE_CONTROL) return POD_DEC_BAD_TYPE;
-  if (f->len < 1 || f->len - 1 > sizeof(c->args)) return POD_DEC_BAD_PAYLOAD;
+  if (f->len < 1 || (size_t)(f->len - 1) > sizeof(c->args)) return POD_DEC_BAD_PAYLOAD;
   c->seq = f->seq;
   c->opcode = f->payload[0];
   c->arg_len = (uint8_t)(f->len - 1);
